+++ src/store/Utils/filtersData.ts
@@ -64,63 +64,6 @@ export type CountriesOptions = "Argentina"| "Australia"| "Austria"| "Belgium"| "
                                 "Singapore"| "Slovakia"| "Slovenia"| "South_Africa"| "South_Korea"| "Sweden"| "Switzerland"| "Taiwan"| "Thailand"| "Turkey"| "UAE"| "Ukraine"| 
                                 "United_Kingdom"| "United_States"| "Venuzuela"; 
                                 
-// export const CountriesObjArray = [
-//   { id: "ar", name: "Argentina" },
-//   { id: "au", name: "Australia" },
-//   { id: "at", name: "Austria" },
-//   { id: "be", name: "Belgium" },
-//   { id: "br", name: "Brazil" },
-//   { id: "bg", name: "Bulgaria" },
-//   { id: "ca", name: "Canada" },
-//   { id: "cn", name: "China" },
-//   { id: "co", name: "Colombia" },
-//   { id: "cu", name: "Cuba" },
-//   { id: "cz", name: "Czech Republic" },
-//   { id: "eg", name: "Egypt" },
-//   { id: "fr", name: "France" },
-//   { id: "de", name: "Germany" },
-//   { id: "gr", name: "Greece" },
-//   { id: "hk", name: "Hong Kong" },
-//   { id: "hu", name: "Hungary" },
-//   { id: "in", name: "India" },
-//   { id: "id", name: "Indonesia" },
-//   { id: "ie", name: "Ireland" },
-//   { id: "il", name: "Israel" },
-//   { id: "it", name: "Italy" },
-//   { id: "jp", name: "Japan" },
-//   { id: "lv", name: "Latvia" },
-//   { id: "lt", name: "Lithuania" },
-//   { id: "my", name: "Malaysia" },
-//   { id: "mx", name: "Mexico" },
-//   { id: "ma", name: "Morocco" },
-//   { id: "nl", name: "Netherlands" },
-//   { id: "nz", name: "New Zealand" },
-//   { id: "ng", name: "Nigeria" },
-//   { id: "no", name: "Norway" },
-//   { id: "ph", name: "Philippines" },
-//   { id: "pl", name: "Poland" },
-//   { id: "pt", name: "Portugal" },
-//   { id: "ro", name: "Romania" },
-//   { id: "ru", name: "Russia" },
-//   { id: "sa", name: "Saudi Arabia" },
-//   { id: "rs", name: "Serbia" },
-//   { id: "sg", name: "Singapore" },
-//   { id: "sk", name: "Slovakia" },
-//   { id: "si", name: "Slovenia" },
-//   { id: "za", name: "South Africa" },
-//   { id: "kr", name: "South Korea" },
-//   { id: "se", name: "Sweden" },
-//   { id: "ch", name: "Switzerland" },
-//   { id: "tw", name: "Taiwan" },
-//   { id: "th", name: "Thailand" },
-//   { id: "tr", name: "Turkey" },
-//   { id: "ae", name: "UAE" },
-//   { id: "ua", name: "Ukraine" },
-//   { id: "gb", name: "United Kingdom" },
-//   { id: "us", name: "United States" },
-//   { id: "us", name: "Venuzuela" },
-// ];
-
 export const Languages = {
   Amharic: "ar",
   German: "de",
